// src/app/enums/table-filter-type.ts
export enum TableFilterType {
	TEXT = 'text',
	NUMBER = 'number',
	DATE = 'date',
	DATE_TIME = 'date_time',
	SELECT = 'select',
	MULTI_SELECT = 'multi_select',
	NONE = 'none',
	SINGLE_DATE = 'single_date',
	STRING_DATE = 'string_date',
	COUNTRY = 'country',
	CURRENCY = 'currency',
	GROUP = 'group',
	EMPLOYEE = 'employee',
	BOOLEAN = 'boolean'
}